Add more tests for base
