+++ tests/test_base.py
@@ -16,142 +16,6 @@ def supress_warnings():
     yield
     warnings.simplefilter("default")
 
-@pytest.mark.parametrize('x,expected', [
-    ([1,2,3], ['1', '2', '3']),
-    (numpy.array([1,2,3]), ['1','2','3']),
-    (1, '1'),
-])
-def test_as_character(x, expected):
-    assert_equal(as_character(x), expected)
-
-@pytest.mark.parametrize('x,format,try_formats,optional,tz,origin,expected', [
-    (["1jan1960", "2jan1960", "31mar1960", "30jul1960"],
-     "%d%b%Y", None, False, 0, None,
-     [datetime.date(1960, 1, 1),
-      datetime.date(1960, 1, 2),
-      datetime.date(1960, 3, 31),
-      datetime.date(1960, 7, 30)]),
-    (["02/27/92", "02/27/92", "01/14/92", "02/28/92", "02/01/92"],
-     "%m/%d/%y", None, False, 0, None,
-     [datetime.date(1992, 2, 27),
-      datetime.date(1992, 2, 27),
-      datetime.date(1992, 1, 14),
-      datetime.date(1992, 2, 28),
-      datetime.date(1992, 2, 1)]),
-    (32768,
-     None, None, False, 0, "1900-01-01",
-     [datetime.date(1989, 9, 19)]),
-    (35981,
-     None, None, False, 0, "1899-12-30",
-     [datetime.date(1998, 7, 5)]),
-    (34519,
-     None, None, False, 0, "1904-01-01",
-     [datetime.date(1998, 7, 5)]),
-    (734373 - 719529,
-     None, None, False, 0, datetime.datetime(1970, 1, 1),
-     [datetime.date(2010, 8, 23)]),
-    ([datetime.date(2010, 4, 13)],
-     None, None, False, 12, None,
-     [datetime.date(2010, 4, 13)]),
-    ([datetime.date(2010, 4, 13)],
-     None, None, False, 12 + 13, None,
-     [datetime.date(2010, 4, 14)]),
-    ([datetime.datetime(2010, 4, 13)],
-     None, None, False, 0 - 10, None,
-     [datetime.date(2010, 4, 12)]),
-])
-def test_as_date(x, format, try_formats, optional, tz, origin, expected):
-    assert_equal(as_date(
-        x,
-        format,
-        try_formats,
-        optional,
-        tz,
-        origin
-    ), expected)
-
-def test_as_date_error():
-    with pytest.raises(ValueError):
-        as_date(1.1)
-
-    with pytest.raises(ValueError):
-        as_date("1990-1-1", "%Y")
-
-    assert as_date("1990-1-1", "Y", optional=True).isna().all()
-
-@pytest.mark.parametrize('x,expected', [
-    ([1,2,3], [1.0, 2.0, 3.0]),
-    (numpy.array(['1','2','3']), [1.0, 2.0, 3.0])
-])
-def test_as_double(x, expected):
-    assert_equal(as_double(x), expected)
-
-@pytest.mark.parametrize('x,expected', [
-    ([1,2,3], Categorical([1,2,3])),
-    (numpy.array(['1','2','3']), Categorical(['1', '2', '3']))
-])
-def test_as_categorical(x, expected):
-    assert_equal(as_categorical(x), expected)
-
-@pytest.mark.parametrize('x,expected', [
-    ([1.0,2.0,3.0], [1,2,3]),
-    (numpy.array(['1','2','3']), [1,2,3]),
-    (factor(['1','2','3']), [0,1,2]),
-])
-def test_as_int(x, expected):
-    assert_equal(as_int(x), expected)
-
-@pytest.mark.parametrize('x,expected', [
-    ([1.0,2.0,3.0], [1,2,3]),
-    (numpy.array(['1','2','3']), [1,2,3])
-])
-def test_as_integer(x, expected):
-    assert_equal(as_integer(x), expected)
-
-@pytest.mark.parametrize('x,expected', [
-    ([1.0,0.0], [True, False]),
-    (numpy.array(['1','0']), [True, False])
-])
-def test_as_logical(x, expected):
-    assert_equal(as_logical(x), expected)
-
-def test_factor():
-    ff = factor(list("statistics"), levels=letters)
-    assert_equal(ff.to_numpy().flatten(), list("statistics"))
-    assert_equal(ff.categories.tolist(), letters)
-
-    ffi = as_integer(ff)
-    assert_equal(ffi, [18,19,0,19,8,18,19,8,2,18])
-
-    ff = factor(ff)
-    assert_equal(ff.categories.tolist(), list('acist'))
-
-    z = factor(LETTERS[-3:], ordered=TRUE)
-    assert "['X' < 'Y' < 'Z']" in str(z)
-
-    x = factor(c(1, 2, NA), exclude=None)
-    assert len(x) == 3
-    # different from R
-    assert_equal(x.categories.tolist(), [1, 2])
-
-    x = factor(z, exclude = "X")
-    assert is_na(x[0])
-    assert len(x.categories) == 2
-
-@pytest.mark.parametrize('x,expected', [
-    (1.4, 2),
-    ([1.1,2.1], [2, 3])
-])
-def test_ceiling(x, expected):
-    assert_equal(ceiling(x), expected)
-
-@pytest.mark.parametrize('x,expected', [
-    (1.4, 1),
-    ([1.1,2.1], [1, 2])
-])
-def test_floor(x, expected):
-    assert_equal(floor(x), expected)
-
 def test_c():
     assert c(1,2,3) == [1,2,3]
     assert c(c(1,2), 3) == [1,2,3]
@@ -163,53 +27,6 @@ def test_rowcolnames():
     df = DataFrame([1,2,3], index=['a', 'b', 'c'])
     assert colnames(df) == [0]
     assert rownames(df) == ['a', 'b', 'c']
-
-def test_cumx():
-    #                                  1,2,3,4,5
-    assert_equal(cumsum(range(1,6)),  [1,3,6,10,15])
-    assert_equal(cumprod(range(1,6)), [1,2,6,24,120])
-    assert_equal(cummin([3,2,1]), [3,2,1])
-    assert_equal(cummax([3,2,1]), [3,3,3])
-
-def test_cut():
-    z = stats.rnorm(10000)
-    tab = table(cut(z, breaks=range(-6, 7)))
-    assert tab.shape == (1, 12)
-    assert tab.columns.tolist() == [
-        Interval(-6, -5, closed='right'),
-        Interval(-5, -4, closed='right'),
-        Interval(-4, -3, closed='right'),
-        Interval(-3, -2, closed='right'),
-        Interval(-2, -1, closed='right'),
-        Interval(-1, 0, closed='right'),
-        Interval(0, 1, closed='right'),
-        Interval(1, 2, closed='right'),
-        Interval(2, 3, closed='right'),
-        Interval(3, 4, closed='right'),
-        Interval(4, 5, closed='right'),
-        Interval(5, 6, closed='right'),
-    ]
-    assert sum(tab.values.flatten()) == 10000
-
-    z = cut([1] * 5, 4)
-    assert_equal(z.to_numpy(), [Interval(0.9995, 1.0, closed='right')] * 5)
-    assert_equal(z.categories.to_list(), [
-        Interval(0.999,  0.9995, closed='right'),
-        Interval(0.9995, 1.0, closed='right'),
-        Interval(1.0,    1.0005, closed='right'),
-        Interval(1.0005, 1.001, closed='right'),
-    ])
-
-    z = stats.rnorm(100)
-    tab = table(cut(z, breaks=[pi/3.0*i for i in range(0,4)]))
-    assert str(tab.columns.tolist()[0]) == '(0.0, 1.05]'
-
-    tab = table(cut(z, breaks=[pi/3.0*i for i in range(0,4)], precision=3))
-    assert str(tab.columns.tolist()[0]) == '(0.0, 1.047]'
-
-    aaa = c(1,2,3,4,5,2,3,4,5,6,7)
-    ct = cut(aaa, 3, precision=3, ordered_result=True)
-    assert str(ct[0]) == '(0.994, 3.0]'
 
 def test_diag():
     assert dim(diag(3)) == (3,3)
@@ -224,20 +41,6 @@ def test_diag():
     x = diag(c(2,1), 4)
     assert_iterable_equal(diag(x), [2,1,2,1])
 
-def test_sample():
-    x = sample(range(1, 13))
-    assert set(x) == set(range(1,13))
-
-    y = sample(x, replace=TRUE)
-    assert len(y) <= 12
-
-    z = sample(c(0,1), 100, replace=TRUE)
-    assert set(z) == {0, 1}
-    assert len(z) == 100
-
-    w = sample('abc', 100, replace=TRUE)
-    assert set(w) == {'a', 'b', 'c'}
-    assert len(z) == 100
 
 def test_table():
     # https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/table
@@ -328,46 +131,3 @@ def test_table_error():
         table(warpbreaks, iris)
     with pytest.raises(ValueError):
         table(warpbreaks.wool, iris)
-
-@pytest.mark.parametrize('x, times, length, each, expected', [
-    (range(4), 2, None, 1, [0,1,2,3]*2),
-    (range(4), 1, None, 2, [0,0,1,1,2,2,3,3]),
-    (range(4), [2]*4, None, 1, [0,0,1,1,2,2,3,3]),
-    (range(4), [2,1]*2, None, 1, [0,0,1,2,2,3]),
-    (range(4), 1, 4, 2, [0,0,1,1]),
-    (range(4), 1, 10, 2, [0,0,1,1,2,2,3,3,0,0]),
-    (range(4), 3, None, 2, [0,0,1,1,2,2,3,3] * 3),
-    (1, 7, None, 1, [1,1,1,1,1,1,1]),
-])
-def test_rep(x, times, length, each, expected):
-    assert_iterable_equal(rep(x, times=times, length=length, each=each), expected)
-
-def test_rep_error():
-    with pytest.raises(ValueError):
-        rep(c(1,2,3), c(1,2))
-    with pytest.raises(ValueError):
-        rep(c(1,2,3), c(1,2,3), each=2)
-
-def test_droplevels():
-    x = factor(c(1,2,3), levels=c(1,2,3,4))
-    x = droplevels(x)
-    assert len(x.categories) == 3
-
-@pytest.mark.parametrize('from_, to, by, length_out, along_with, expect', [
-    (0, 1, None, 11, None, [.0,.1,.2,.3,.4,.5,.6,.7,.8,.9,1.]),
-    ([4,2,1,9,3]*4, None, None, None, None, range(1, 21)),
-    (1, 9, 2, None, None, [1,3,5,7,9]),
-    (1, 9, 2, None, None, [1,3,5,7,9]),
-    (1, 6, 3, None, None, [1, 4]),
-    (1.575, 5.125, 0.05, None, None, [
-        1.575, 1.625, 1.675, 1.725, 1.775, 1.825, 1.875, 1.925, 1.975, 2.025, 2.075, 2.125,
-        2.175, 2.225, 2.275, 2.325, 2.375, 2.425, 2.475, 2.525, 2.575, 2.625, 2.675, 2.725,
-        2.775, 2.825, 2.875, 2.925, 2.975, 3.025, 3.075, 3.125, 3.175, 3.225, 3.275, 3.325,
-        3.375, 3.425, 3.475, 3.525, 3.575, 3.625, 3.675, 3.725, 3.775, 3.825, 3.875, 3.925,
-        3.975, 4.025, 4.075, 4.125, 4.175, 4.225, 4.275, 4.325, 4.375, 4.425, 4.475, 4.525,
-        4.575, 4.625, 4.675, 4.725, 4.775, 4.825, 4.875, 4.925, 4.975, 5.025, 5.075, 5.125,
-    ]),
-    (17, None, None, None, None, range(1,18))
-])
-def test_seq(from_, to, by, length_out, along_with, expect):
-    assert_iterable_equal(seq(from_, to, by, length_out, along_with), expect, approx=True)
